Added game icon support
This is the coolest thing I've ever done.

Essentially, I parse the input and clear out all unwanted characters, but I also store the locations of each string that is an icon in the game files. I then substitute a special character from a esoteric or practically unused Unicode sets for the string I want to replace, and then I iterate through the string *again* and this time append each image in each location. There probably is a better way to simplify the process, but this allows for easy editing of both encoding and decoding dictionaries.

In the future, I plan to more cohesively declare each dictionary, and make the encoding dictionary rely on the decoding dictionary for values or something similar, that way all encoding characters can be edited in one place.