Fixed output cleaning
The output is now cleaned up a lot better and is far more readable. It includes a portion of text I'm not sure I want to include as it's not in the game output, but it does also make the text a bit more readable, so I will potentially make a settings option for this or something.