Fixed output to be cleaner
Added checkpoint icon and fixed the output to be a bit cleaner than before; I removed the previously left-in option of having the code titles of each segment (like ManTalkShop001) so that the output is truly what the game sees. I intend to add button support as soon as I can.